Removing trailing whitespace
Trailing whitespace on lines is always bad; it screws up various text editors,
adds noise to diff outputs and is all-around evil. The biggest problem is that
it creates commit noise: it shows some lines as changed when they in fact
haven't been, only whitespace was added/removed at the end.

Some references:

http://codeimpossible.com/2012/04/02/Trailing-whitespace-is-evil-Don-t-commit-evil-into-your-repo-/
http://programmers.stackexchange.com/questions/121555/why-is-trailing-whitespace-a-big-deal