Add package.json so that Pegged can be used as a DUB dependency.
Adds a minimal package.json file. The library will be compiled as a
source library together with the host project.

Trying to build directly using "dub" currently fails because no main()
is defined. In the future this may be changed to instead build a
static library, but dub needs to be extended for that slightly.

The package file here is basically the same as in #109, but removes
all build type specific flags (dub manages those itself using e.g.
--build=unittest).