Relocated examples to a separate folder and usage of subpackages
So we avoid making the pegged dub package too heavy with untested code.
The most interesting code, from the point of view to be reused by
grammar composition, are shared with dub subpackages.

Having all the examples on a different folder, each one with a dub.json,
allow us to execute examples unit tests on the CI tests.